requirements.txt
fastapi
uvicorn
google-genai
noisereduce
numpy
pyaudio
openai-whisper
python-dotenv
gTTS
langdetect
pydub
